pyproject.toml
[project]
name = "prasaarit-upload-service"
version = "0.1.0"
description = "Upload service for prasaarit"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "aws-lambda-powertools>=3.26.0",
    "fastapi[standard]>=0.135.2",
    "mangum>=0.21.0",
    "pydantic>=2.12.5",
    "ulid-py>=1.1.0",
]

[dependency-groups]
dev = [
    "boto3>=1.42.73",
    "boto3-stubs[s3]>=1.42.73",
    "pre-commit>=4.5.1",
    "types-boto3[essential]>=1.42.73",
]
lint = [
    "bandit>=1.9.4",
    "black>=26.3.1",
    "flake8>=7.3.0",
    "flake8-pyproject>=1.2.4",
    "isort>=8.0.1",
]

[tool.fastapi]
entrypoint = "src.main:app"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
